+++ weblate/appsettings.py
@@ -19,7 +19,7 @@
 #
 
 from django.conf import settings
-from weblate.trans.util import get_script_name
+from weblate.trans.scripts import get_script_name
 import os
 
 
@@ -71,6 +71,9 @@ DEMO_SERVER = get('DEMO_SERVER', False)
 
 # Enable remote hooks
 ENABLE_HOOKS = get('ENABLE_HOOKS', True)
+
+# Enable sharing
+ENABLE_SHARING = get('ENABLE_SHARING', True)
 
 # Whether to run hooks in background
 BACKGROUND_HOOKS = get('BACKGROUND_HOOKS', True)