+++ weblate/trans/models/unit.py
@@ -27,7 +27,6 @@ from django.contrib import messages
 from django.core.cache import cache
 import traceback
 from weblate.trans.checks import CHECKS
-from weblate.trans.models.translation import Translation
 from weblate.trans.models.source import Source
 from weblate.trans.search import update_index_unit, fulltext_search, more_like
 
@@ -299,7 +298,7 @@ class UnitManager(models.Manager):
 
 
 class Unit(models.Model):
-    translation = models.ForeignKey(Translation)
+    translation = models.ForeignKey('Translation')
     checksum = models.CharField(max_length=40, db_index=True)
     contentsum = models.CharField(max_length=40, db_index=True)
     location = models.TextField(default='', blank=True)
@@ -673,7 +672,7 @@ class Unit(models.Model):
 
         # Update checks if content or fuzzy flag has changed
         if not same_content or not same_state:
-            self.check(same_state, same_content, force_insert)
+            self.run_checks(same_state, same_content, force_insert)
 
         # Update fulltext index if content has changed or this is a new unit
         if force_insert or not same_content:
@@ -853,7 +852,7 @@ class Unit(models.Model):
 
         return checks_to_run, cleanup_checks
 
-    def check(self, same_state=True, same_content=True, is_new=False):
+    def run_checks(self, same_state=True, same_content=True, is_new=False):
         """
         Updates checks for this unit.
         """