+++ weblate/trans/views/api.py
@@ -31,6 +31,7 @@ from weblate.trans.util import get_site_url
 import json
 import weblate
 import threading
+import re
 
 
 BITBUCKET_GIT_REPOS = (
@@ -49,6 +50,17 @@ GITHUB_REPOS = (
     'https://github.com/%(owner)s/%(slug)s.git',
     'git@github.com:%(owner)s/%(slug)s.git',
 )
+
+HOOK_HANDLERS = {}
+
+
+def register_hook(handler):
+    """
+    Registers hook handler.
+    """
+    name = handler.__name__.split('_')[0]
+    HOOK_HANDLERS[name] = handler
+    return handler
 
 
 def perform_update(obj):
@@ -91,8 +103,9 @@ def git_service_hook(request, service):
     '''
     Shared code between Git service hooks.
 
-    Currently used for bitbucket_hook and github_hook, but should be usable for
-    hook from other Git services (Google Code, custom coded sites, etc.) too.
+    Currently used for bitbucket_hook, github_hook and gitlab_hook, but should
+    be usable for other Git services (Google Code, custom coded sites, etc.)
+    too.
     '''
     # Check for enabled hooks
     if appsettings.ENABLE_HOOKS:
@@ -106,18 +119,20 @@ def git_service_hook(request, service):
 
     # Check if we got payload
     try:
-        data = json.loads(request.POST['payload'])
+        # GitLab sends json as application/json
+        if request.META['CONTENT_TYPE'] == 'application/json':
+            data = json.loads(request.body)
+        # Bitbucket and GitHub sends json as x-www-form-data
+        else:
+            data = json.loads(request.POST['payload'])
     except (ValueError, KeyError):
         return HttpResponseBadRequest('Could not parse JSON payload!')
 
     # Get service helper
-    if service == 'github':
-        hook_helper = github_hook_helper
-    elif service == 'bitbucket':
-        hook_helper = bitbucket_hook_helper
-    else:
-        weblate.logger.error('service %s, not supported', service)
+    if service not in HOOK_HANDLERS:
+        weblate.logger.error('service %s is not supported', service)
         return HttpResponseBadRequest('invalid service')
+    hook_helper = HOOK_HANDLERS[service]
 
     # Send the request data to the service handler.
     try:
@@ -152,7 +167,7 @@ def git_service_hook(request, service):
     return HttpResponse('update triggered')
 
 
-@csrf_exempt
+@register_hook
 def bitbucket_hook_helper(data):
     '''
     API to handle commit hooks from Bitbucket.
@@ -185,7 +200,7 @@ def bitbucket_hook_helper(data):
     }
 
 
-@csrf_exempt
+@register_hook
 def github_hook_helper(data):
     '''
     API to handle commit hooks from GitHub.
@@ -193,13 +208,32 @@ def github_hook_helper(data):
     # Parse owner, branch and repository name
     owner = data['repository']['owner']['name']
     slug = data['repository']['name']
-    branch = data['ref'].split('/')[-1]
+    branch = re.sub(r'^refs/heads/', '', data['ref'])
 
     # Construct possible repository URLs
     repos = [repo % {'owner': owner, 'slug': slug} for repo in GITHUB_REPOS]
 
     return {
         'service_long_name': 'GitHub',
+        'repos': repos,
+        'branch': branch,
+    }
+
+
+@register_hook
+def gitlab_hook_helper(data):
+    '''
+    API to handle commit hooks from GitLab.
+    '''
+    ssh_url = data['repository']['url']
+    http_url = '.'.join((data['repository']['homepage'], 'git'))
+    branch = re.sub(r'^refs/heads/', '', data['ref'])
+
+    # Construct possible repository URLs
+    repos = [ssh_url, http_url]
+
+    return {
+        'service_long_name': 'GitLab',
         'repos': repos,
         'branch': branch,
     }
@@ -226,7 +260,7 @@ def export_stats(request, project, subproject):
 
     try:
         indent = int(request.GET['indent'])
-    except:
+    except (ValueError, KeyError):
         indent = None
 
     response = []